# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin role required"
          },
          "409": {
            "description": "Sku already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new product (Admin Only)",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by category",
            "schema": {
              "example": "Electronics",
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by product name",
            "schema": {
              "example": "headphones",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of products with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get all products with optional filters",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": " Get product by id",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "SKu already exists"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a product (Admin Only)",
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "400": {
            "description": "Cannot delete product in active orders"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete product (Admin Only) ",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}/stock": {
      "patch": {
        "operationId": "ProductsController_updateStock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "number",
                    "description": "Stock adjustment ( positive to add, negative to subtract) ",
                    "example": 10
                  }
                },
                "required": [
                  "quantity"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stock updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Insufficient stock"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update product stock (Admin Only)",
        "tags": [
          "Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Wireless Headphones",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Prodcut description",
            "example": "High-quality wireless headphones with noise cancellation"
          },
          "price": {
            "type": "number",
            "description": "Product price in usd",
            "example": 99.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "Stock quantity",
            "example": 100,
            "minimum": 0
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping Unit (Sku) -unique identifier",
            "example": "WH-001",
            "maxLength": 50
          },
          "imageUrl": {
            "type": "string",
            "description": "Product image url",
            "example": "https://example.com/image.jpg"
          },
          "categoryId": {
            "type": "string",
            "description": "Product category",
            "example": "Electronics"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether product is active and available for purchase",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "price",
          "stock",
          "sku",
          "categoryId"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Product ID",
            "example": "46545646sds-4584s68sd-4654684sd"
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Wireless Headphone"
          },
          "description": {
            "type": "object",
            "description": "Product description",
            "example": "High quality wireless headphones",
            "nullable": true
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 99.99
          },
          "stock": {
            "type": "number",
            "description": "Product stock",
            "example": 100
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping Unit",
            "example": "WH-001"
          },
          "imageUrl": {
            "type": "object",
            "description": "Product image url",
            "example": "https://example.com/image.jpg"
          },
          "category": {
            "type": "object",
            "description": "Product category",
            "example": "Electronics"
          },
          "isActive": {
            "type": "boolean",
            "description": "Product availability status",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "last update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "price",
          "stock",
          "sku",
          "imageUrl",
          "category",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Wireless Headphones",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Prodcut description",
            "example": "High-quality wireless headphones with noise cancellation"
          },
          "price": {
            "type": "number",
            "description": "Product price in usd",
            "example": 99.99,
            "minimum": 0
          },
          "stock": {
            "type": "number",
            "description": "Stock quantity",
            "example": 100,
            "minimum": 0
          },
          "sku": {
            "type": "string",
            "description": "Stock keeping Unit (Sku) -unique identifier",
            "example": "WH-001",
            "maxLength": 50
          },
          "imageUrl": {
            "type": "string",
            "description": "Product image url",
            "example": "https://example.com/image.jpg"
          },
          "categoryId": {
            "type": "string",
            "description": "Product category",
            "example": "Electronics"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether product is active and available for purchase",
            "example": true,
            "default": true
          }
        }
      },
...
    }
  }
}